Remove neons plugin.
Neons is no longer needed, remove neons plugin and HPDatabaseType.kNeons
and HPDatabaseType.kNeonsAndRadon. enum values kRadon and kNoDatabase
remain.

Further streamlining of radon access: parameter information was previously
fetched from many different plugins. Now it is centered to two places:
compiled_plugin_base::SetParams and luatool::SetParam. These two functions
fetch all necessary parameter id information etc from database, so it's no
longer necessary to re-fetch them later (like when writing file to disk).

+++ himan-lib/source/json_parser.cpp
@@ -632,7 +632,7 @@ raw_time GetLatestOriginDateTime(const shared_ptr<configuration> conf, const str
 
 	auto r = GET_PLUGIN(radon);
 
-	auto latestFromDatabase = r->RadonDB().GetLatestTime(static_cast<int>(sourceProducer.Id()), "", offset);
+	auto latestFromDatabase = r->RadonDB().GetLatestTime(sourceProducer.Id(), "", offset);
 
 	if (!latestFromDatabase.empty())
 	{
